Extract Report class to own module
